숫자에 콤마 추가, sfx 새로 적용, 게임 오버 시 남은 행성들이 모두 정리가 안되는 문제 수정 등

+++ Client/Assets/Scripts/UI/Prefabs/UIGameoverPopup.cs
@@ -29,7 +29,7 @@ public class UIGameoverPopup : UI
         });
     }
 
-    public void Bind(long score, long prevHighestScore, long curHighestScore)
+    public void Bind(int score, int prevHighestScore, int curHighestScore)
     {
         _scoreText.text = $"현재 점수 : {score}";
         _bestScoreText.text = $"최고 점수 : {curHighestScore}";
